--- Shoplive_iOS_ParkJiHo/SceneDelegate.swift
@@ -1,23 +1,0 @@
-//
-//  SceneDelegate.swift
-//  Shoplive_iOS_ParkJiHo
-//
-//  Created by jiho park on 5/20/24.
-//
-
-import UIKit
-
-class SceneDelegate: UIResponder, UIWindowSceneDelegate {
-
-    var window: UIWindow?
-
-    func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
-        guard let windowScene = (scene as? UIWindowScene) else { return }
-        
-        let vc = ViewController()
-        let window = UIWindow(windowScene: windowScene)
-        self.window = window
-        window.makeKeyAndVisible()
-        window.rootViewController = vc
-    }
-}
